backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
langchain
langchain-groq
langchain-community
beautifulsoup4
sentence-transformers
faiss-cpu
requests
gunicorn